feat(config): make websocket_delay a float

+++ avabot/config.py
@@ -12,7 +12,14 @@ class AvaApiConfig:
 
 
 class SlackConfig:
-    def __init__(self, bot_id: str, bot_name: str, api_token: str, whitelist_channels: str, websocket_delay: int) -> None:
+    def __init__(
+            self,
+            bot_id: str,
+            bot_name: str,
+            api_token: str,
+            whitelist_channels: str,
+            websocket_delay: float
+    ) -> None:
         self.bot_id = bot_id
         self.bot_name = bot_name
         self.api_token = api_token
@@ -35,7 +42,7 @@ def _load_unsafe() -> Dict:
 
         'SLACK_API_TOKEN': os.environ['SLACK_API_TOKEN'],
         'SLACK_WHITELIST_CHANNELS': os.environ['SLACK_WHITELIST_CHANNELS'],
-        'SLACK_WEBSOCKET_DELAY': int(os.environ.get('SLACK_WEBSOCKET_DELAY', 1)),
+        'SLACK_WEBSOCKET_DELAY': float(os.environ.get('SLACK_WEBSOCKET_DELAY', 1)),
         'SLACK_BOT_ID': os.environ['SLACK_BOT_ID'],
         'SLACK_BOT_NAME': os.environ.get('SLACK_BOT_NAME', 'ava'),
     }
